Changed function placements and mogrify for pg INSERT
