+++ ee/api/app.py
@@ -16,6 +16,7 @@ from chalicelib.utils import events_queue
 from routers import core, core_dynamic, ee, saml
 from routers.crons import core_crons
 from routers.crons import core_dynamic_crons
+from routers.crons import ee_crons
 from routers.subs import dashboard, insights, metrics, v1_api_ee
 from routers.subs import v1_api
 
@@ -85,7 +86,7 @@ async def startup():
     await events_queue.init()
     app.schedule.start()
 
-    for job in core_crons.cron_jobs + core_dynamic_crons.cron_jobs + traces.cron_jobs:
+    for job in core_crons.cron_jobs + core_dynamic_crons.cron_jobs + traces.cron_jobs + ee_crons.ee_cron_jobs:
         app.schedule.add_job(id=job["func"].__name__, **job)
 
     ap_logger.info(">Scheduled jobs:")
