+++ ee/scripts/schema/db/init_dbs/postgresql/1.10.0/1.10.0.sql
@@ -7,13 +7,13 @@ $$ LANGUAGE sql IMMUTABLE;
 
 CREATE TABLE IF NOT EXISTS frontend_signals
 (
-    project_id    bigint                                         NOT NULL,
-    user_id       text                                           NOT NULL,
-    timestamp     bigint                                         NOT NULL,
-    action        text                                           NOT NULL,
-    source        text                                           NOT NULL,
-    category      text                                           NOT NULL,
-    data          json
+    project_id bigint  NOT NULL,
+    user_id    integer NOT NULL references users (user_id) ON DELETE CASCADE,
+    timestamp  bigint  NOT NULL,
+    action     text    NOT NULL,
+    source     text    NOT NULL,
+    category   text    NOT NULL,
+    data       jsonb
 );
 CREATE INDEX IF NOT EXISTS frontend_signals_user_id_idx ON frontend_signals (user_id);
